requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
Django==5.0.3
djangorestframework==3.15.1
drf-yasg==1.21.7
h11==0.14.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
pycparser==2.21
PySocks==1.7.1
pytest==8.1.1
pytz==2024.1
PyYAML==6.0.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0

marshmallow~=3.21.1
Faker~=24.4.0
assertpy~=1.1
requests~=2.31.0